restructured choccy_chip package
