+++ choccy_chip/src/lib.rs
@@ -6,8 +6,6 @@
 pub mod prelude;
 /// Emulator API
 pub mod emulator;
-/// Graphics API
-pub mod graphics;
 // /// Input API
 // pub mod input;
 // /// Audio API
